# [FIX] product: speedup _name_search for dynamic variants templates